requirements.txt
robotframework~=6.0.2
robotframework-seleniumlibrary==5.1.3
selenium
percy-selenium
